+++ tileserver/maptiles/mbtilesgenerator_postgres.go
@@ -235,19 +235,20 @@ func (self *TileDbPostgresql) rowExists(query string) bool {
 }
 
 // MetaDataHandler gets metadata from database.
-func (self *TileDbPostgresql) MetaDataHandler(lyr string) map[string]string {
+func (self *TileDbPostgresql) MetaDataHandler(lyr string) (map[string]string, error) {
+	metadata := make(map[string]string)
 	rows, err := self.db.Query("SELECT name, value FROM metadata WHERE layer_name=$1", lyr)
 	if nil != err {
 		Ligneous.Error(err)
+		return metadata, err
 	}
-	metadata := make(map[string]string)
 	for rows.Next() {
 		var name string
 		var value string
 		rows.Scan(&name, &value)
 		metadata[name] = value
 	}
-	return metadata
+	return metadata, nil
 }
 
 // GetTileLayers get metadata for all tilelayers.
@@ -261,7 +262,12 @@ func (self *TileDbPostgresql) GetTileLayers() (map[string]map[string]string, err
 	for rows.Next() {
 		var layer_name string
 		rows.Scan(&layer_name)
-		layers[layer_name] = self.MetaDataHandler(layer_name)
+		metadata, err := self.MetaDataHandler(layer_name)
+		if nil != err {
+			Ligneous.Error(err)
+			return layers, err
+		}
+		layers[layer_name] = metadata
 	}
 	return layers, nil
 }