+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Change Log
 ## [Unreleased]
 
 
+## [0.1.6] - 2017-04-07
+### Changed
+ - Read tilelayers from database
+
+
 ## [0.1.5] - 2017-04-03
 ### Added
  - Global variables